+++ CHANGES.rst
@@ -11,7 +11,7 @@ Unreleased
 
 * When calling :meth:`.Session.request`, we add the key-value pair
 ``"api_type": "json"`` to the ``json`` parameter, if it is a ``dict``.
-* :class:`.ScriptAuthorizer` has a new parameter ``otp_function`` that supplies
+* :class:`.ScriptAuthorizer` has a new parameter ``two_factor_callback `` that supplies
   OTPs (One-Time Passcodes) every hour. 
 
 1.3.0 (2020-04-23)